gpu: nvgpu: Cast unsigned arith to signed for abs()
Before passing in an unsigned value to abs() cast the result
of the subtraction to signed.

In Linux this happens automatically but on non-Linux platforms
abs() does not necessarily do this. clang flags this case as
a pointless operation: abs(x) obviously must equal x for any
unsigned x.

This change should hopefully preserve the Linux behavior but
avoid the compiler warning from clang.

JIRA NVGPU-525

Reviewed-on: https://git-master.nvidia.com/r/1724093
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/gm20b/clk_gm20b.c
@@ -204,7 +204,7 @@ static int clk_config_pll(struct clk_gk20a *clk, struct pll *pll,
 				if (vco_f >= min_vco_f && vco_f <= max_vco_f) {
 					lwv = (vco_f + (nvgpu_pl_to_div(pl) / 2))
 						/ nvgpu_pl_to_div(pl);
-					delta = abs(lwv - target_clk_f);
+					delta = abs((s32)(lwv - target_clk_f));
 
 					if (delta < best_delta) {
 						best_delta = delta;
